renamed constants to swerve constants
